snap-bootstrap: refactor partition creation
Refactor filesystem creation, deployment and mounting to act on a single
partition instead of a slice of partitions.

Based on Michael Vogt's work in PR #7723.

+++ cmd/snap-bootstrap/partition/mkfs.go
@@ -24,12 +24,14 @@ import (
 	"github.com/snapcore/snapd/gadget"
 )
 
-func MakeFilesystems(created []DeviceStructure) error {
-	for _, part := range created {
-		if part.VolumeStructure.Filesystem != "" {
-			if err := mkfs(part.Node, part.VolumeStructure.Label, part.VolumeStructure.Filesystem); err != nil {
-				return err
-			}
+func MakeFilesystem(part DeviceStructure) error {
+	if part.VolumeStructure == nil {
+		return fmt.Errorf("cannot use incomplete device %v", part.Node)
+	}
+
+	if part.VolumeStructure.Filesystem != "" {
+		if err := mkfs(part.Node, part.VolumeStructure.Label, part.VolumeStructure.Filesystem); err != nil {
+			return err
 		}
 	}
 	return nil
